requirements.txt
PySide6~=6.9.0
cython
opencv-python~=4.11.0.86
numpy~=2.2.6
pillow~=11.2.1
setuptools~=68.2.0
regex~=2024.11.6
uiautomator2
pyinstaller